Deregistration of global objects at framework shutdown time
The change ensures that the framework unregisters all the global fixtures and
configuration when it is shut down. This prevents the global fixtures/configuration
to deregister themselves at destruction time, which requires the framework to be
still existing, and for which the order of destruction cannot be ensured during
the unwinding of static/global objects.